requirements.txt
absl-py
attrs
jax
jaxlib
numpy
numpyro
matplotlib
sklearn
tensorflow-addons
tensorflow-federated==0.20.0
tensorflow-probability==0.16.0
cloudpickle
contexttimer
optax
dm-haiku
blackjax
fedjax
jaxopt
Deprecated
hydra-joblib-launcher
imageio